migrate: move ProgrammaticMigration to separate file
